+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [FUTURE] (2026-05-29)
 ## bug fixes
 * AON RoadPathfinder. uses time_columns and acces_time to do the pathfinder and the car_los time col
+* distribute_commons_on_links: only put boardings on first links and alightings on last link (when there is multiple link merge together to create a common_link, not between trips) 
 
 ## [3.6.0] (2026-05-25)
 ## Features